requirements.txt
code_generator @ file:///Users/abakunov/Desktop/Barcode_analysis
ghostscript==0.7
numpy==2.1.3
pillow==11.0.0
setuptools==75.6.0
treepoem==3.25.0
